fix(NumberInput): pass the evt and dir to events
onChange and onClick will now receive the events used to trigger them.
This is a regression fix.

They will also receive the direction as a second argument if the arrows
are used.

Fixes the tests as they were returning false positives when checking on
spies. The spies were not reset between test runs so passing it in one
test means it passed in the rest.

Fixes #369