Update Dropdown Role to ListBox
The role presentation attached to dropdown isn't considered a form element. Therefore when partnered with a <label> we get an accessibility error:
```
The id "identifier" referenced by the 'for' attribute of a label element is not a valid form input element or an element with id "identifier" is either null or does not exist.
```

Changing the role to Listbox makes the dropdown act as if it is a select. However each child will need to also have the role option so screen readers know which children are selectable. See #

References:
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_listbox_role
https://www.w3.org/TR/wai-aria/roles#listitem