feat(DatePicker): allow specifying the DOM node the calendar dropdown is inserted into
Flatpickr does not change the position of calendar dropdown once it's created,
even after the element containing the triggering element (`<input>`) is scrolled.

Flatpickr allows specifying the DOM node the calendar dropdown is inserted into,
which would work as a way to alleviate such scrolling issue,
for example, by specifying a DOM element being scrolled along with the triggering element.

This change adds a prop that will be passed along to Flatpickr as such DOM node.

Fixes #357.